+++ FS22_ExtendedGameInfoDisplay/tempinfo.lua
@@ -13,7 +13,7 @@ TempInfo = {}
 ---@param overwrittenFunc function
 ---@param isVisible boolean
 function TempInfo:gameinfodisplay__setTemperatureVisible(overwrittenFunc, isVisible)
-    overwrittenFunc(self, true)
+    overwrittenFunc(self, ExtendedGameInfoDisplayGui.settings.temperaturVisible)
 end
 
 ---Overwritten GameInfoDisplay:updateTemperature()